PYT-811 Add plugin wrapper calls to upgrade decorator (revert)
Reviewed at: http://reviews.delphix.com/r/53587/
