+++ build.gradle
@@ -8,7 +8,7 @@ plugins {
 }
 
 subprojects {
-  version = "1.0.1-internal-004"
+  version = "1.0.1-internal-003"
 }
 
 def binDir = "${rootProject.projectDir}/bin"
