+++ platform/lock.dev-requirements.txt
@@ -1,18 +1,18 @@
 -i https://artifactory.delphix.com/artifactory/api/pypi/delphix-virtual-pypi/simple/
 atomicwrites==1.3.0
-attrs==19.3.0
+attrs==19.1.0
 configparser==4.0.2 ; python_version < '3'
-contextlib2==0.6.0.post1 ; python_version < '3'
-funcsigs==1.0.2 ; python_version < '3.0'
+contextlib2==0.5.5 ; python_version < '3'
+funcsigs==1.0.2 ; python_version < '3.3'
 importlib-metadata==0.23 ; python_version < '3.8'
 mock==3.0.5
 more-itertools==5.0.0 ; python_version <= '2.7'
-packaging==19.2
-pathlib2==2.3.5 ; python_version == '3.4.*' or python_version < '3'
+packaging==19.1
+pathlib2==2.3.4 ; python_version == '3.4.*' or python_version < '3'
 pluggy==0.13.0
 py==1.8.0
 pyparsing==2.4.2
-pytest==4.6.6
+pytest==4.6.5
 scandir==1.10.0 ; python_version < '3.5'
 six==1.12.0
 wcwidth==0.1.7
