+++ platform/lock.requirements.txt
@@ -1,5 +1,5 @@
 -i https://artifactory.delphix.com/artifactory/api/pypi/delphix-virtual-pypi/simple/
-./../common/build/python-dist/dvp-common-1.0.1-internal-004.tar.gz
+./../common/build/python-dist/dvp-common-1.0.1-internal-003.tar.gz
 enum34==1.1.6 ; python_version < '3.4'
 protobuf==3.6.1
 six==1.12.0
