+++ platform/src/main/python/dlpx/virtualization/platform/__init__.py
@@ -5,8 +5,4 @@
 __path__ = __import__('pkgutil').extend_path(__path__, __name__)
 
 from dlpx.virtualization.platform._plugin_classes import *
-from dlpx.virtualization.platform._discovery import *
-from dlpx.virtualization.platform._linked import *
-from dlpx.virtualization.platform._upgrade import *
-from dlpx.virtualization.platform._virtual import *
 from dlpx.virtualization.platform._plugin import *
