+++ tools/lock.dev-requirements.txt
@@ -1,34 +1,34 @@
 -i https://artifactory.delphix.com/artifactory/api/pypi/delphix-virtual-pypi/simple/
-./../common/build/python-dist/dvp-common-1.0.1-internal-004.tar.gz
-./../libs/build/python-dist/dvp-libs-1.0.1-internal-004.tar.gz
-./../platform/build/python-dist/dvp-platform-1.0.1-internal-004.tar.gz
+./../common/build/python-dist/dvp-common-1.0.1-internal-003.tar.gz
+./../libs/build/python-dist/dvp-libs-1.0.1-internal-003.tar.gz
+./../platform/build/python-dist/dvp-platform-1.0.1-internal-003.tar.gz
 atomicwrites==1.3.0
-attrs==19.3.0
+attrs==19.1.0
 backports.functools-lru-cache==1.5 ; python_version < '3.2'
-configparser==4.0.2 ; python_version < '3'
-contextlib2==0.6.0.post1 ; python_version < '3'
+configparser==4.0.2 ; python_version < '3.2'
+contextlib2==0.5.5 ; python_version < '3'
 coverage==4.5.4
 entrypoints==0.3
 enum34==1.1.6
-flake8==3.7.9
-funcsigs==1.0.2 ; python_version < '3.3'
-functools32==3.2.3.post2 ; python_version < '3.2'
+flake8==3.7.8
+funcsigs==1.0.2 ; python_version < '3.0'
+functools32==3.2.3.post2 ; python_version < '3'
 futures==3.3.0 ; python_version < '3.2'
-httpretty==0.9.7
-importlib-metadata==0.23
+httpretty==0.9.6
+importlib-metadata==0.23 ; python_version < '3.8'
 isort==4.3.21
 mccabe==0.6.1
 mock==3.0.5
-more-itertools==5.0.0
-packaging==19.2
-pathlib2==2.3.5 ; python_version == '3.4.*' or python_version < '3'
+more-itertools==5.0.0 ; python_version <= '2.7'
+packaging==19.1
+pathlib2==2.3.4 ; python_version < '3.6'
 pluggy==0.13.0
 py==1.8.0
 pycodestyle==2.5.0
 pyflakes==2.1.1
 pyparsing==2.4.2
-pytest-cov==2.8.1
-pytest==4.6.6
+pytest-cov==2.7.1
+pytest==4.6.5
 scandir==1.10.0 ; python_version < '3.5'
 six==1.12.0
 typing==3.7.4.1 ; python_version < '3.5'
