+++ tools/lock.requirements.txt
@@ -1,31 +1,25 @@
 -i https://artifactory.delphix.com/artifactory/api/pypi/delphix-virtual-pypi/simple/
-attrs==19.3.0
+attrs==19.1.0
 certifi==2019.9.11
 chardet==3.0.4
 click-configfile==0.2.3
 click==7.0
-configparser==4.0.2 ; python_version < '3'
-contextlib2==0.6.0.post1 ; python_version < '3'
+configparser==4.0.2 ; python_version < '3.2'
 entrypoints==0.3
 enum34==1.1.6
-flake8==3.7.9
-functools32==3.2.3.post2 ; python_version < '3.2'
+flake8==3.7.8
+functools32==3.2.3.post2 ; python_version < '3'
 idna==2.8
-importlib-metadata==0.23
-jinja2==2.10.3
-jsonschema==3.1.1
+jinja2==2.10.1
+jsonschema==3.0.2
 markupsafe==1.1.1
 mccabe==0.6.1
-more-itertools==5.0.0
-pathlib2==2.3.5 ; python_version == '3.4.*' or python_version < '3'
 protobuf==3.6.1
 pycodestyle==2.5.0
 pyflakes==2.1.1
-pyrsistent==0.15.5
+pyrsistent==0.15.4
 pyyaml==5.1.2
 requests==2.22.0
-scandir==1.10.0 ; python_version < '3.5'
 six==1.12.0
 typing==3.7.4.1 ; python_version < '3.5'
-urllib3==1.25.6
-zipp==0.6.0
+urllib3==1.25.3
